+++ system/user.go
@@ -20,7 +20,7 @@ import (
 	"syscall"
 	"time"
 
-	"pkg.re/essentialkaos/ek.v8/env"
+	"pkg.re/essentialkaos/ek.v9/env"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //