+++ changelog.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### v9.0.0
+
+* Package `args` renamed to `options` (_incompatible changes_)
+* `[fmtutil/table]` Added new package for rendering data as a table
+* `[fmtutil]` Added support of separator symbol configuration
+* `[usage]` Improved output about a newer version
+* `[usage]` Increased code coverage (0.0% → 100%)
+* `[usage]` Code refactoring
+
+---
+
 ### v8.0.3
 
 * `[usage]` Improved options and commands info rendering